Restructure the comparison folder
